pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vulca"
version = "0.19.0"
description = "Agent-native cultural art SDK for image generation, L1-L5 evaluation, layer editing, visual discovery, and cultural traditions. Agents drive creative decisions; Vulca provides hands + eyes."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Yu Haorui", email = "yuhaorui48@gmail.com" },
]
keywords = ["ai", "art", "evaluation", "cultural", "vlm", "pipeline", "creation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "litellm>=1.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "Pillow>=10.0",
]

[project.optional-dependencies]
# Minimal: evaluate + create (no heavy deps)
core = []

# Layer system: split, composite, export
layers = [
    "Pillow>=10.0",
    "cairosvg>=2.7",
    "numpy>=1.24",
]

# Algorithmic tools: cv2-based analysis
tools = [
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "Pillow>=10.0",
]

# SAM segmentation (optional, heavy)
# Supports two pipelines pending convergence (issue #N+5 — v0.18.0):
#   - SAM v1 (segment-anything) — used by scripts/claude_orchestrated_pipeline.py
#     via sam_model_registry["vit_l"] at hardcoded /tmp/sam_vit_l.pth.
#   - SAM v2 (sam2) — used by src/vulca/layers/sam.py via facebook/sam2.1-hiera-small.
# timm>=1.0 required because transformers 4.49 timm_wrapper needs
# timm.data.ImageNetInfo (introduced in timm 0.9+); torchscale pin to timm==0.6.13
# is tolerated (torchscale not on orchestrated path).
# SAM v1 checkpoint /tmp/sam_vit_l.pth is NOT auto-downloaded; see docs/INSTALL-RECIPE.md.
sam = [
    "torch>=2.0",
    "numpy>=1.24",
    "Pillow>=10.0",
    "segment-anything>=1.0",
    "timm>=1.0.0",
]

# SAM3 text-prompted segmentation (optional, heavy, GPU-only)
sam3 = [
    "transformers>=4.50",
    "torch>=2.0",
    "numpy>=1.24",
    "Pillow>=10.0",
]

# Background removal (optional)
rembg = [
    "rembg>=2.0",
]

# Scout: semantic search
scout = [
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.0",
]

# MCP server
mcp = [
    "fastmcp>=3.0",
]

# Full layer system with all optional extras
layers-full = [
    "Pillow>=10.0",
    "cairosvg>=2.7",
    "numpy>=1.24",
    "rembg>=2.0",
    "transformers>=4.50",
    "torch>=2.0",
]

# Full installation (layers + tools, excluding heavy ML deps)
all = [
    "vulca[layers,tools,mcp,scout]",
]

# Development
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.20",
    "ruff>=0.6",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
# Conservative floor: catch real bugs only (GitHub Actions recommended set).
# Expand to E/W/I in v0.17.12+ when legacy code is cleaned incrementally.
select = ["E9", "F63", "F7", "F82"]

[project.urls]
Homepage = "https://vulcaart.art"
Documentation = "https://github.com/vulca-org/vulca"
Repository = "https://github.com/vulca-org/vulca"
Issues = "https://github.com/vulca-org/vulca/issues"

[project.scripts]
vulca = "vulca.cli:main"
vulca-mcp = "vulca.mcp_server:mcp.run"

[tool.hatch.build.targets.wheel]
packages = ["src/vulca"]

[tool.hatch.build]
artifacts = ["src/vulca/cultural/data/**"]

[tool.hatch.build.targets.sdist]
# Whitelist approach — ship only what's needed to build from source.
# Previous exclude-list missed root-level model weights (*.pt, *.pth, .ts),
# generated PNGs, and screenshots, bloating sdist past PyPI's 100MB limit.
# Wheel already ships only src/vulca/, so sdist parity is this include set.
include = [
    "/src/vulca",
    "/pyproject.toml",
    "/README.md",
    "/CHANGELOG.md",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "l4_local: requires local model weights (DINO + SAM); skipped in CI via pytest.importorskip('torch')",
    "real_provider: requires real cloud provider credentials (e.g. OpenAI, Gemini, etc); skipped in CI",
    "local_provider: requires local provider stack (ComfyUI + Ollama); skipped in CI",
]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_return_any = false
warn_unused_configs = true

# Modules with pre-existing type errors. New files are checked by default.
# Remove entries as you fix their errors to tighten coverage incrementally.
[[tool.mypy.overrides]]
module = [
    "vulca._engine",
    "vulca._vlm",
    "vulca.cli",
    "vulca.cultural.loader",
    "vulca.digestion.dream",
    "vulca.digestion.evolver",
    "vulca.digestion.preloader",
    "vulca.digestion.signals",
    "vulca.digestion.trajectory",
    "vulca.layers.blend",
    "vulca.layers.composite",
    "vulca.layers.export",
    "vulca.layers.matting",
    "vulca.layers.plan_prompt",
    "vulca.layers.redraw",
    "vulca.layers.sam3",
    "vulca.layers.split",
    "vulca.layers.transform",
    "vulca.layers.vlm_mask",
    "vulca.mcp_server",
    "vulca.pipeline.engine",
    "vulca.pipeline.nodes.evaluate",
    "vulca.pipeline.nodes.layer_generate",
    "vulca.pipeline.residuals",
    "vulca.providers.gemini",
    "vulca.stats",
    "vulca.studio.brief",
    "vulca.studio.phases.inpaint",
    "vulca.studio.phases.intent",
    "vulca.studio.session",
    "vulca.tools.adapters.cli",
    "vulca.tools.adapters.mcp",
    "vulca.tools.adapters.pipeline",
    "vulca.tools.adapters.sdk",
    "vulca.tools.cultural.brushstroke",
    "vulca.tools.cultural.color_gamut",
    "vulca.tools.cultural.composition",
    "vulca.tools.cultural.whitespace",
    "vulca.tools.filters.color_correct",
]
ignore_errors = true
